Modify : 결제하기 전송 함수 로직 변경
